Plug small memory leak when a thread exits. Allow new threads to inherit scheduling policy. Use long instead of int in thread_exit, thanks to the ports maintainers for the heads up.
svn path=/trunk/thread/; revision=5179

+++ src/thread/thread.h
@@ -162,7 +162,7 @@ void thread_rwlock_rlock_c(rwlock_t *rwlock, int line, char *file);
 void thread_rwlock_wlock_c(rwlock_t *rwlock, int line, char *file);
 void thread_rwlock_unlock_c(rwlock_t *rwlock, int line, char *file);
 void thread_rwlock_destroy(rwlock_t *rwlock);
-void thread_exit_c(int val, int line, char *file);
+void thread_exit_c(long val, int line, char *file);
 
 /* sleeping */
 void thread_sleep(unsigned long len);
